[SMALLFIX] Move data structure into tachyon.collections
