+++ common/src/test/java/tachyon/underfs/UnderFileSystemTest.java
@@ -20,7 +20,7 @@ import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
 
-import tachyon.Pair;
+import tachyon.collections.Pair;
 import tachyon.TachyonURI;
 import tachyon.conf.TachyonConf;
