+++ servers/src/main/java/tachyon/master/file/FileSystemMaster.java
@@ -36,8 +36,8 @@ import com.google.common.collect.Sets;
 import tachyon.Constants;
 import tachyon.HeartbeatExecutor;
 import tachyon.HeartbeatThread;
-import tachyon.Pair;
-import tachyon.PrefixList;
+import tachyon.collections.Pair;
+import tachyon.collections.PrefixList;
 import tachyon.StorageLevelAlias;
 import tachyon.TachyonURI;
 import tachyon.conf.TachyonConf;
